v8.9.0-alpha
In the Analyser: Added a note to the importer recommending users import no more than 500 entries at a time to ensure stability.
In the Analyser: Implemented data pagination in the data table to improve performance with large datasets.
In the Analyser: Users can now select which columns to display for each collection, and these preferences are saved.