v9.79.0-alpha
RecipeManagr
Added a 'Resting Time' field to the Add/Edit form and displayed it in the Recipe Details modal.
Introduced 'Flavor Profile' checkboxes (Sweet/Savoury) to allow better categorization of recipes.
Updated the 'Date Discovered' field to automatically default to the current date when adding a new recipe.
Enhanced the Recipe Details modal to explicitly show the Source Name and link it to the Source URL if available.